move KeyStore to new file
