+++ tff_aggregations/channels/channel_grid.py
@@ -5,14 +5,11 @@ from tensorflow_federated.python.common_libs import py_typecheck
 from tensorflow_federated.python.core.impl.types import placement_literals
 
 from tff_aggregations.channels import channel
-
-PlacementPair = Tuple[
-    placement_literals.PlacementLiteral,
-    placement_literals.PlacementLiteral]
+from tff_aggregations import utils
 
 @dataclass
 class ChannelGrid:
-  _channel_dict: Dict[PlacementPair, channel.Channel]
+  _channel_dict: Dict[utils.PlacementPair, channel.Channel]
   requires_setup: bool = True
 
   async def setup_channels(self, strategy):
@@ -26,7 +23,7 @@ class ChannelGrid:
         self._channel_dict[placement_pair] = channel
       self.requires_setup = False
 
-  def __getitem__(self, placements: PlacementPair):
+  def __getitem__(self, placements: utils.PlacementPair):
     py_typecheck.check_type(placements, tuple)
     py_typecheck.check_len(placements, 2)
     sorted_placements = sorted(placements, key=lambda p: p.uri)
